chore: 스케줄러 이동 domain → application
