+++ src/main/java/com/hhplus/concert/application/scheduler/QueueScheduler.java
@@ -1,16 +1,19 @@
-package com.hhplus.concert.domain.reservation;
+package com.hhplus.concert.application.scheduler;
 
+import com.hhplus.concert.domain.queue.QueueService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 
 @Component
 @RequiredArgsConstructor
-public class ReservationScheduler {
-    private final ReservationService reservationService;
+public class QueueScheduler {
+
+    private final QueueService queueService;
 
     @Scheduled(initialDelay = 60000, fixedDelay = 60000)
-    public void releaseSeatsOnPaymentExpired() {
-        reservationService.releaseSeatsOnPaymentExpired();
+    public void activate() {
+        queueService.activate();
     }
+
 }
